Pass to 8Bits sub code type, little endian (for x86 and arm performance improvement)

+++ general/base/ProtocolParsing.h
@@ -139,13 +139,13 @@ protected:
 protected:
     //have message without reply
     virtual void parseMessage(const quint8 &mainCodeType,const char * const data,const int &size) = 0;
-    virtual void parseFullMessage(const quint8 &mainCodeType,const quint16 &subCodeType,const char * const data,const int &size) = 0;
+    virtual void parseFullMessage(const quint8 &mainCodeType,const quint8 &subCodeType,const char * const data,const int &size) = 0;
     //have query with reply
     virtual void parseQuery(const quint8 &mainCodeType,const quint8 &queryNumber,const char * const data,const int &size) = 0;
-    virtual void parseFullQuery(const quint8 &mainCodeType,const quint16 &subCodeType,const quint8 &queryNumber,const char * const data,const int &size) = 0;
+    virtual void parseFullQuery(const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber,const char * const data,const int &size) = 0;
     //send reply
     virtual void parseReplyData(const quint8 &mainCodeType,const quint8 &queryNumber,const char * const data,const int &size) = 0;
-    virtual void parseFullReplyData(const quint8 &mainCodeType,const quint16 &subCodeType,const quint8 &queryNumber,const char * const data,const int &size) = 0;
+    virtual void parseFullReplyData(const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber,const char * const data,const int &size) = 0;
 
     virtual void reset();
 private:
@@ -163,20 +163,20 @@ private:
 public:
     //reply to the query
     QHash<quint8,quint8> waitedReply_mainCodeType;
-    QHash<quint8,quint16> waitedReply_subCodeType;
+    QHash<quint8,quint8> waitedReply_subCodeType;
     #ifndef EPOLLCATCHCHALLENGERSERVERNOCOMPRESSION
     QSet<quint8> replyOutputCompression;
     #endif
     QHash<quint8,quint16> replyOutputSize;
 public:
     void newOutputQuery(const quint8 &mainCodeType,const quint8 &queryNumber);
-    void newFullOutputQuery(const quint8 &mainCodeType,const quint16 &subCodeType,const quint8 &queryNumber);
+    void newFullOutputQuery(const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber);
     //send message without reply
     bool packOutcommingData(const quint8 &mainCodeType,const char * const data,const int &size);
-    bool packFullOutcommingData(const quint8 &mainCodeType,const quint16 &subCodeType,const char * const data,const int &size);
+    bool packFullOutcommingData(const quint8 &mainCodeType,const quint8 &subCodeType,const char * const data,const int &size);
     //send query with reply
     bool packOutcommingQuery(const quint8 &mainCodeType,const quint8 &queryNumber,const char * const data,const int &size);
-    bool packFullOutcommingQuery(const quint8 &mainCodeType,const quint16 &subCodeType,const quint8 &queryNumber,const char * const data,const int &size);
+    bool packFullOutcommingQuery(const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber,const char * const data,const int &size);
     //send reply
     bool postReplyData(const quint8 &queryNumber, const char * const data,const int &size);
 
@@ -193,7 +193,7 @@ public:
             const bool &isClient,
             #endif
             char *buffer,
-            const quint8 &mainCodeType,const quint16 &subCodeType,const char * const data,const int &size);
+            const quint8 &mainCodeType,const quint8 &subCodeType,const char * const data,const int &size);
     //send query with reply
     static int computeOutcommingQuery(
             #ifndef CATCHCHALLENGERSERVERDROPIFCLENT
@@ -206,7 +206,7 @@ public:
             const bool &isClient,
             #endif
             char *buffer,
-            const quint8 &mainCodeType,const quint16 &subCodeType,const quint8 &queryNumber,const char * const data,const int &size);
+            const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber,const char * const data,const int &size);
     //send reply
     int computeReplyData(char *dataBuffer, const quint8 &queryNumber, const char * const data, const int &size);
     //compression
@@ -219,14 +219,14 @@ private:
 
     // for data
     quint8 mainCodeType;
-    quint32 subCodeType;
+    quint8 subCodeType;
     quint8 queryNumber;
     static QByteArray lzmaCompress(QByteArray data);
     static QByteArray lzmaUncompress(QByteArray data);
     static const quint16 sizeHeaderNullquint16;
 public:
     virtual void storeInputQuery(const quint8 &mainCodeType,const quint8 &queryNumber);
-    virtual void storeFullInputQuery(const quint8 &mainCodeType,const quint16 &subCodeType,const quint8 &queryNumber);
+    virtual void storeFullInputQuery(const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber);
 protected:
     //reply to the query
     #ifdef CATCHCHALLENGER_EXTRA_CHECK
@@ -264,7 +264,7 @@ public:
     quint64 getRXSize() const;
     #endif
     void storeInputQuery(const quint8 &mainCodeType,const quint8 &queryNumber);
-    void storeFullInputQuery(const quint8 &mainCodeType,const quint16 &subCodeType,const quint8 &queryNumber);
+    void storeFullInputQuery(const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber);
 
     void closeSocket();
 protected:
